+++ libs/assembling/assembler/src/lib/combiner/combiner.interface.ts
@@ -1,5 +1,5 @@
 import { TreeToken } from '@idl/parsing/syntax-tree';
-import { TokenName } from '@idl/tokenizer';
+import { TOKEN_NAMES, TokenName } from '@idl/tokenizer';
 
 /**
  * Track the strings by line
@@ -14,6 +14,8 @@ export interface IStringsByLine {
 export interface ICombinerRecursionOptions {
   /** Are we within a comment block? */
   commentBlock: boolean;
+  /** indicates if we are a hanging line with custom indent */
+  hangingIndentStart: number | undefined;
   /** If we should ignore line continuations that we encounter and not add indents */
   ignoreLineContinuation: boolean;
   /** Indent level for joining strings */
@@ -27,3 +29,15 @@ export interface ICombinerRecursionOptions {
   /** The current parents */
   tokenParent: TreeToken<TokenName> | undefined;
 }
+
+/** Function tokens for customized property behavior */
+export const HANGING_ROUTINES: { [key: string]: boolean } = {};
+HANGING_ROUTINES[TOKEN_NAMES.CALL_FUNCTION] = true;
+HANGING_ROUTINES[TOKEN_NAMES.CALL_FUNCTION_METHOD] = true;
+HANGING_ROUTINES[TOKEN_NAMES.CALL_PROCEDURE] = true;
+HANGING_ROUTINES[TOKEN_NAMES.CALL_PROCEDURE_METHOD] = true;
+HANGING_ROUTINES[TOKEN_NAMES.ROUTINE_NAME] = true;
+HANGING_ROUTINES[TOKEN_NAMES.ROUTINE_METHOD_NAME] = true;
+HANGING_ROUTINES[TOKEN_NAMES.PARENTHESES] = true;
+HANGING_ROUTINES[TOKEN_NAMES.BRACKET] = true;
+HANGING_ROUTINES[TOKEN_NAMES.STRUCTURE] = true;