+++ libs/assembling/assembler/src/lib/combiner/combiner.ts
@@ -12,6 +12,7 @@ import { NonBasicTokenNames, TOKEN_NAMES, TokenName } from '@idl/tokenizer';
 import { GetNewLine } from '../helpers/get-new-line';
 import { Stringify } from '../helpers/stringify';
 import {
+  HANGING_ROUTINES,
   ICombinerRecursionOptions,
   IStringsByLine,
 } from './combiner.interface';
@@ -42,7 +43,8 @@ function _Recursor<T extends FormatterType>(
   let line = GetNewLine(
     recurse.indentLevel,
     options.tabWidth,
-    recurse.tokenParent
+    recurse.tokenParent,
+    recurse.hangingIndentStart
   );
 
   /** Store the last line from our tokens */
@@ -111,7 +113,8 @@ function _Recursor<T extends FormatterType>(
         line = GetNewLine(
           recurse.indentLevel,
           options.tabWidth,
-          recurse.tokenParent
+          recurse.tokenParent,
+          recurse.hangingIndentStart
         );
         strings[lastLine] = line;
       }
@@ -161,6 +164,20 @@ function _Recursor<T extends FormatterType>(
       recurse.tokenBefore = branch as TreeToken<TokenName>;
       recurse.tokenParent = branch as TreeToken<TokenName>;
 
+      // track last hang number
+      let lastHang: number;
+
+      // see if we should have hanging indent
+      if (options.hangingIndent && tree[i].name in HANGING_ROUTINES) {
+        // if we had a previous value, save it so we can reset when we return
+        if (recurse.hangingIndentStart) {
+          lastHang = recurse.hangingIndentStart;
+        }
+
+        // set to the end of the start of our branch
+        recurse.hangingIndentStart = tree[i].pos[1] + tree[i].pos[2];
+      }
+
       // process our children
       _Recursor(branch.kids, options, recurse, strings);
 
@@ -195,10 +212,12 @@ function _Recursor<T extends FormatterType>(
               closeAdd = GetNewLine(
                 recurse.indentLevel - indentOffset,
                 options.tabWidth,
-                recurse.tokenParent
+                recurse.tokenParent,
+                recurse.hangingIndentStart
               );
               closeAdd.push(toAdd);
-              // closeAdd.push(toAdd);
+
+              // save strings for our new line
               strings[closeLine] = closeAdd;
             } else {
               // if existing strings, concat to last entry to avoid excess strings
@@ -209,10 +228,21 @@ function _Recursor<T extends FormatterType>(
           line = GetNewLine(
             recurse.indentLevel - indentOffset,
             options.tabWidth,
-            recurse.tokenParent
+            recurse.tokenParent,
+            recurse.hangingIndentStart
           );
           line.push(toAdd);
           strings[closeLine] = line;
+        }
+      }
+
+      // check if we need to reset our hanging indent
+      if (options.hangingIndent && tree[i].name in HANGING_ROUTINES) {
+        if (lastHang) {
+          recurse.hangingIndentStart = lastHang;
+          lastHang = undefined;
+        } else {
+          recurse.hangingIndentStart = undefined;
         }
       }
 
@@ -262,6 +292,7 @@ export function Combiner<T extends FormatterType>(
 
   // create our recursion options
   const recursion: ICombinerRecursionOptions = {
+    hangingIndentStart: undefined,
     indentLevel: 0,
     lineContinuation: false,
     ignoreLineContinuation: false,