+++ libs/tests/config-resolution/src/lib/bad-file.spec.ts
@@ -28,6 +28,7 @@ describe(`[auto generated] Parse invalid config files`, () => {
     // verify results
     expect({
       formatter: 'fiddle',
+      hangingIndent: false,
       tabWidth: 2,
       eol: 'lf',
       style: {